backend/requirements.txt
agora_token_builder==1.0.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
chardet==5.2.0
charset-normalizer==3.4.1
constantly==23.10.4
coverage==7.8.0
cryptography==44.0.2
cssselect2==0.8.0
daphne==4.2.0
defusedxml==0.7.1
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-oauth-toolkit==3.0.1
django-request==1.7.0
django-test-plus==2.2.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
fonttools==4.57.0
fpdf==1.7.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.10.0
googleapis-common-protos==1.69.2
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
jwcrypto==1.5.6
oauthlib==3.2.2
packaging==25.0
pillow==11.1.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg==3.2.6
psycopg-binary==3.2.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydyf==0.11.0
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
pyphen==0.17.2
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-http-client==3.3.7
python3-openid==3.2.0
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sendgrid==6.11.0
service-identity==24.2.0
setuptools==80.7.1
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
stripe==7.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
weasyprint==65.1
webencodings==0.5.1
websockets==15.0.1
zope.interface==7.2
zopfli==0.2.3.post1
